requirements.txt
streamlit>=1.28.0
bcrypt>=4.0.1
pillow>=10.0.0
reportlab>=4.0.0
google-generativeai>=0.3.0
openai>=1.0.0
